+++ src/apscheduler/executors/pool.py
@@ -5,6 +5,11 @@ from concurrent.futures.process import BrokenProcessPool
 
 from apscheduler.executors.base import BaseExecutor, run_job
 
+SHUTDOWN_MESSAGES = {
+    "cannot schedule new futures after shutdown",
+    "cannot schedule new futures after interpreter shutdown",
+}
+
 
 class BasePoolExecutor(BaseExecutor):
     @abstractmethod
@@ -24,9 +29,20 @@ class BasePoolExecutor(BaseExecutor):
             else:
                 self._run_job_success(job.id, f.result())
 
-        f = self._pool.submit(
-            run_job, job, job._jobstore_alias, run_times, self._logger.name
-        )
+        try:
+            f = self._pool.submit(
+                run_job, job, job._jobstore_alias, run_times, self._logger.name
+            )
+        except RuntimeError as e:
+            # There is a possible race condition between shutdown being invoked and the job being submitted to pool.
+            # This helps with not polluting logs with RuntimeError traces.
+            if str(e) in SHUTDOWN_MESSAGES:
+                self._logger.warning(
+                    "Executor has been shut down, skipping job %s: %s", job.id, e
+                )
+                return
+            else:
+                raise  # Re-raise unexpected RuntimeErrors
         f.add_done_callback(callback)
 
     def shutdown(self, wait=True):